+++ networking/src/main/kotlin/com/ppereira/networking/model/RepositoryResponse.kt
@@ -9,8 +9,8 @@ class RepositoryResponse(
     @Json(name = "id") val id: Long,
     @Json(name = "name") val name: String,
     @Json(name = "description") val description: String,
-    @Json(name = "forks_count") val numberOfForks: Long,
-    @Json(name = "stargazers_count") val numberOfWatchers: Long,
-    @Json(name = "open_issues_count") val numberOfOpenIssues: Long,
+    @Json(name = "forks_count") val numberOfForks: Int,
+    @Json(name = "stargazers_count") val numberOfWatchers: Int,
+    @Json(name = "open_issues_count") val numberOfOpenIssues: Int,
     @Json(name = "owner") val owner: PersonResponse
 ) : Parcelable
